infra: add missing SSM permission;
Now that dev & prd share same cluster we need to permit
the cluster to access both paths to the SSM secrets.

This is where DB_PASSWORD, TOKEN, and other sensitive
app secrets are stored securely, and they are accessed at
init time each time a new version of the app is deployed.